feat(topic): Add authorization for topic deletion and update
Enhance topic management by adding authorization checks
for both updating and deleting topics.  Only the topic
creator can now perform these actions.  Unauthorized
attempts will result in an appropriate exception.

+++ src/main/java/br/com/soupaulodev/forumhub/modules/topic/usecase/DeleteTopicUseCase.java
@@ -1,6 +1,7 @@
 package br.com.soupaulodev.forumhub.modules.topic.usecase;
 
 import br.com.soupaulodev.forumhub.modules.exception.usecase.TopicNotFoundException;
+import br.com.soupaulodev.forumhub.modules.exception.usecase.UnauthorizedException;
 import br.com.soupaulodev.forumhub.modules.topic.entity.TopicEntity;
 import br.com.soupaulodev.forumhub.modules.topic.repository.TopicRepository;
 import org.springframework.stereotype.Service;
@@ -30,12 +31,17 @@ public class DeleteTopicUseCase {
      * Executes the use case to delete a topic.
      *
      * @param id the unique identifier of the topic to be deleted
+     * @param authenticatedUserId the authenticated user's unique identifier
      * @throws TopicNotFoundException if the topic specified by the id does not exist
+     * @throws UnauthorizedException if the authenticated user is not the creator of the topic
      */
-    public void execute(UUID id) {
-
+    public void execute(UUID id, UUID authenticatedUserId) {
         TopicEntity topicDB = topicRepository.findById(id)
                 .orElseThrow(TopicNotFoundException::new);
+
+        if (!topicDB.getCreator().getId().equals(authenticatedUserId)) {
+            throw new UnauthorizedException("You are not allowed to delete this topic.");
+        }
 
         topicRepository.delete(topicDB);
     }
